:recycle: cleanup JavaScript
